Ready Up On Network & Spawning
Ready Up Manager is not fully on the network, it gets spawned in when you join the network locker scene. The levers now check the hinge state changed instead of the value of the height of the levers. The levers subscribe to the spawned in ready up manager. Leaving the room should now vacate the tubes.