Added 3rd person Spectator Camera
There is now a third person camera that follows the player around and you can spectate the player. We don't get to see in first person what the player is doing so I will implement to switch to first person next. Scrolling with the mouse wheel will zoom in/out.